% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\docType{class}
\name{Base-class}
\alias{Base-class}
\alias{.Base}
\alias{Base}
\title{Constructor for Base}
\usage{
Base(name, count)
}
\arguments{
\item{name}{A length-one character vector, for instance, 'paul', 'amy'}

\item{count}{A length-one numeric vector}
}
\value{
An object of class Base
}
\description{
Constructor for Base
}
\section{Slots}{

\describe{
\item{\code{name}}{a length-one character vector}

\item{\code{count}}{integer a length-one numeric vector}
}}

\references{
\url{https://en.wikipedia.org/wiki/Abstract_type}
}
