% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derived.R
\docType{class}
\name{Derived-class}
\alias{Derived-class}
\alias{.Derived}
\alias{Derived}
\title{Class Derived}
\usage{
Derived(name, count, height)
}
\arguments{
\item{name}{a length-one character vector}

\item{count}{a length-one numeric vector}

\item{height}{a length-one numeric vector}
}
\value{
an object of class Derived
}
\description{
Class Derived

Constructor for Derived
}
\section{Slots}{

\describe{
\item{\code{height}}{a length-one numeric vector}
}}

