% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\docType{methods}
\name{getCount,Base-method}
\alias{getCount,Base-method}
\alias{getCount}
\title{Retrieve the object's count}
\usage{
\S4method{getCount}{Base}(obj)
}
\arguments{
\item{obj}{An object of class Base}
}
\value{
The current count of the Base object
}
\description{
Retrieve the object's count
}
\examples{
# Create a Base object
base.01 <- Base(name="paul", count=42)
getCount(base.01)

}
