% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derived.R
\docType{methods}
\name{getHeight,Derived-method}
\alias{getHeight,Derived-method}
\alias{getHeight}
\title{Retrieve the height of the object}
\usage{
\S4method{getHeight}{Derived}(obj)
}
\arguments{
\item{obj}{An object of class Base}
}
\value{
The height of the Base object
}
\description{
Retrieve the height of the object
}
\examples{
# Create a Derived object
derived.01 <- Derived(name="paul", count=42, height=65)
getHeight(derived.01)

}
