% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\docType{methods}
\name{getName,Base-method}
\alias{getName,Base-method}
\alias{getName}
\title{Retrieve the name of the object}
\usage{
\S4method{getName}{Base}(obj)
}
\arguments{
\item{obj}{An object of class Base}
}
\value{
The name of the Base object
}
\description{
Retrieve the name of the object
}
\examples{
# Create a Base object
base.01 <- Base(name="paul", count=42)
getName(base.01)

}
